re-org:
1. move SimpleTokenizer to examples
2. move LR to classification, HashingTF and StandardScaler to feature, CV and ParamGridBuilder to tuning
3. define SimpleTransformer
