+++ mllib/src/main/scala/org/apache/spark/ml/param/params.scala
@@ -94,7 +94,7 @@ case class ParamPair[T](param: Param[T], value: T)
  * Trait for components that take parameters. This also provides an internal param map to store
  * parameter values attached to the instance.
  */
-trait Params extends Identifiable {
+trait Params extends Identifiable with Serializable {
 
   /** Returns all params. */
   def params: Array[Param[_]] = {
@@ -296,88 +296,4 @@ object ParamMap {
   }
 }
 
-/**
- * Builder for a param grid used in grid search.
- */
-class ParamGridBuilder {
 
-  private val paramGrid = mutable.Map.empty[Param[_], Iterable[_]]
-
-  /**
-   * Builds base on parameters with fixed values.
-   */
-  def baseOn(paramMap: ParamMap): this.type = {
-    baseOn(paramMap.toSeq: _*)
-    this
-  }
-
-  /**
-   * Builds base on parameters with fixed values.
-   */
-  @varargs
-  def baseOn(paramPairs: ParamPair[_]*): this.type = {
-    paramPairs.foreach { p =>
-      addGrid(p.param.asInstanceOf[Param[Any]], Seq(p.value))
-    }
-    this
-  }
-
-  /**
-   * Adds a param with multiple values (overwrites if the input param exists).
-   */
-  def addGrid[T](param: Param[T], values: Iterable[T]): this.type = {
-    paramGrid.put(param, values)
-    this
-  }
-
-  // specialized versions of addMulti for Java.
-
-  /**
-   * Adds a double param with multiple values.
-   */
-  def addGrid(param: DoubleParam, values: Array[Double]): this.type = {
-    addGrid[Double](param, values)
-  }
-
-  /**
-   * Adds a int param with multiple values.
-   */
-  def addGrid(param: IntParam, values: Array[Int]): this.type = {
-    addGrid[Int](param, values)
-  }
-
-  /**
-   * Adds a float param with multiple values.
-   */
-  def addGrid(param: FloatParam, values: Array[Float]): this.type = {
-    addGrid[Float](param, values)
-  }
-
-  /**
-   * Adds a long param with multiple values.
-   */
-  def addGrid(param: LongParam, values: Array[Long]): this.type = {
-    addGrid[Long](param, values)
-  }
-
-  /**
-   * Adds a boolean param with true and false.
-   */
-  def addGrid(param: BooleanParam): this.type = {
-    addGrid[Boolean](param, Array(true, false))
-  }
-
-  /**
-   * Builds and returns all combinations of parameters specified by the param grid.
-   */
-  def build(): Array[ParamMap] = {
-    var paramMaps = Array(new ParamMap)
-    paramGrid.foreach { case (param, values) =>
-      val newParamMaps = values.flatMap { v =>
-        paramMaps.map(_.copy.put(param.asInstanceOf[Param[Any]], v))
-      }
-      paramMaps = newParamMaps.toArray
-    }
-    paramMaps
-  }
-}
