+++ mllib/src/test/java/org/apache/spark/ml/classification/JavaLogisticRegressionSuite.java
@@ -15,11 +15,9 @@
  * limitations under the License.
  */
 
-package org.apache.spark.ml.example;
+package org.apache.spark.ml.classification;
 
 import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.List;
 
 import org.junit.After;
 import org.junit.Before;
@@ -30,8 +28,12 @@ import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.spark.ml.Pipeline;
 import org.apache.spark.ml.PipelineModel;
 import org.apache.spark.ml.PipelineStage;
-import org.apache.spark.ml.param.ParamGridBuilder;
+import org.apache.spark.ml.evaluation.BinaryClassificationEvaluator;
+import org.apache.spark.ml.feature.StandardScaler;
 import org.apache.spark.ml.param.ParamMap;
+import org.apache.spark.ml.tuning.CrossValidator;
+import org.apache.spark.ml.tuning.CrossValidatorModel;
+import org.apache.spark.ml.tuning.ParamGridBuilder;
 import org.apache.spark.mllib.regression.LabeledPoint;
 import org.apache.spark.mllib.util.MLUtils;
 import org.apache.spark.sql.api.java.JavaSQLContext;
@@ -119,40 +121,6 @@ public class JavaLogisticRegressionSuite implements Serializable {
     PipelineModel model = pipeline.fit(dataset);
     model.transform(dataset).registerTempTable("prediction");
     JavaSchemaRDD predictions = jsql.sql("SELECT label, score, prediction FROM prediction");
-    for (Row r: predictions.collect()) {
-      System.out.println(r);
-    }
-  }
-
-  @Test
-  public void textClassificationPipeline() {
-    List<LabeledDocument> localTraining = new ArrayList<LabeledDocument>();
-    localTraining.add(new LabeledDocument(0L, "a b c d e spark", 1.0));
-    localTraining.add(new LabeledDocument(1L, "b d", 0.0));
-    localTraining.add(new LabeledDocument(2L, "spark f g h", 1.0));
-    localTraining.add(new LabeledDocument(3L, "hadoop mapreduce", 0.0));
-    JavaSchemaRDD training =
-      jsql.applySchema(jsc.parallelize(localTraining), LabeledDocument.class);
-    Tokenizer tokenizer = new Tokenizer()
-      .setInputCol("text")
-      .setOutputCol("words");
-    HashingTF hashingTF = new HashingTF()
-      .setInputCol(tokenizer.getOutputCol())
-      .setOutputCol("features");
-    LogisticRegression lr = new LogisticRegression()
-      .setMaxIter(10);
-    Pipeline pipeline = new Pipeline()
-      .setStages(new PipelineStage[] {tokenizer, hashingTF, lr});
-    PipelineModel model = pipeline.fit(training);
-    List<Document> localTest = new ArrayList<Document>();
-    localTest.add(new Document(0L, "a b c d e spark"));
-    localTest.add(new Document(1L, "b d"));
-    localTest.add(new Document(2L, "spark f g h"));
-    localTest.add(new Document(3L, "hadoop mapreduce"));
-    JavaSchemaRDD test =
-      jsql.applySchema(jsc.parallelize(localTest), Document.class);
-    model.transform(test).registerAsTable("prediction");
-    JavaSchemaRDD predictions = jsql.sql("SELECT id, text, prediction, score FROM prediction");
     for (Row r: predictions.collect()) {
       System.out.println(r);
     }
